+++ spec/item_repository_spec.rb
@@ -1,22 +1,20 @@
-# frozen_string_literal: true
-
 require 'rspec'
 require './lib/item'
 require './lib/item_repository'
 require './lib/file_io'
-require './data/mock_data'
+require './data/mockable'
 
 describe ItemRepository do
   describe '#initialize' do
     it 'exists' do
-      allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(MockData.items_as_hashes)
+      allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(ItemMocks.items_as_hashes)
       item_repository = ItemRepository.new('fake.csv')
 
       expect(item_repository).is_a? ItemRepository
     end
 
     it 'has an items array' do
-      allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(MockData.items_as_hashes)
+      allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(ItemMocks.items_as_hashes)
       item_repository = ItemRepository.new('fake.csv')
 
       expect(item_repository.items).is_a? Array
@@ -25,7 +23,7 @@ describe ItemRepository do
 
   describe '#all' do
     it 'returns a list of all items' do
-      allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(MockData.items_as_hashes)
+      allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(ItemMocks.items_as_hashes)
       item_repository = ItemRepository.new('fake.csv')
 
       expect(item_repository.all.length).to eq 10
@@ -35,7 +33,7 @@ describe ItemRepository do
 
   describe '#find_by_id' do
     it 'returns nil if no item has the specified id' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -43,7 +41,7 @@ describe ItemRepository do
     end
 
     it 'returns the item with the specified id' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -54,7 +52,7 @@ describe ItemRepository do
 
   describe '#find_by_name' do
     it 'returns nil if no item has name specified' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -62,7 +60,7 @@ describe ItemRepository do
     end
 
     it 'returns the item with the specified name' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -73,7 +71,7 @@ describe ItemRepository do
 
   describe '#find_all_with_description' do
     it 'returns empty array if description does not match' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -81,7 +79,7 @@ describe ItemRepository do
     end
 
     it 'returns array of items with matching descriptions' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -91,8 +89,8 @@ describe ItemRepository do
 
   describe '#find_all_by_price' do
     it 'returns an empty array if no items match price' do
-      details = MockData.items_as_hashes(unit_price: 10.00)
-      mock_data = MockData.items_as_mocks(self, details)
+      details = ItemMocks.items_as_hashes(unit_price: 10.00)
+      mock_data = ItemMocks.items_as_mocks(self, details)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -100,8 +98,8 @@ describe ItemRepository do
     end
 
     it 'returns array of items that match specified price' do
-      details = MockData.items_as_hashes(unit_price: 25.00)
-      mock_data = MockData.items_as_mocks(self, details)
+      details = ItemMocks.items_as_hashes(unit_price: 25.00)
+      mock_data = ItemMocks.items_as_mocks(self, details)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -111,8 +109,8 @@ describe ItemRepository do
 
   describe '#find_all_by_price_in_range(range)' do
     it 'returns empty array if no items in price range' do
-      details = MockData.items_as_hashes(unit_price: 8.00)
-      mock_data = MockData.items_as_mocks(self, details)
+      details = ItemMocks.items_as_hashes(unit_price: 8.00)
+      mock_data = ItemMocks.items_as_mocks(self, details)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -120,8 +118,8 @@ describe ItemRepository do
     end
 
     it 'returns array of all items in price range' do
-      details = MockData.items_as_hashes(unit_price: 25.00)
-      mock_data = MockData.items_as_mocks(self, details)
+      details = ItemMocks.items_as_hashes(unit_price: 25.00)
+      mock_data = ItemMocks.items_as_mocks(self, details)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -131,7 +129,7 @@ describe ItemRepository do
 
   describe '#find_all_by_merchant_id' do
     it 'returns an empty array if no items with merchant_id' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -139,7 +137,7 @@ describe ItemRepository do
     end
 
     it 'returns all items with merchant_id' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -149,7 +147,7 @@ describe ItemRepository do
 
   describe '#create' do
     it 'creates an Item class object' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -166,7 +164,7 @@ describe ItemRepository do
     end
 
     it 'creates an Item with a new id' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -188,7 +186,7 @@ describe ItemRepository do
 
   describe '#update' do
     it 'updates the item with new attributes' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
       new_item = {
@@ -218,7 +216,7 @@ describe ItemRepository do
     end
 
     it 'updates the item with new time' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
       new_item = {
@@ -247,7 +245,7 @@ describe ItemRepository do
 
   describe '#delete' do
     it 'deletes the object at specified id' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
 
@@ -270,7 +268,7 @@ describe ItemRepository do
     end
 
     it 'does not delete anything if no item at id' do
-      mock_data = MockData.items_as_mocks(self)
+      mock_data = ItemMocks.items_as_mocks(self)
       allow_any_instance_of(ItemRepository).to receive(:create_items).and_return(mock_data)
       item_repository = ItemRepository.new('fake.csv')
