+++ spec/item_spec.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 require 'rspec'
 require './lib/item'
